splitting scopes and better structure
