+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/function/TransformScpoe.kt
@@ -1,0 +1,13 @@
+package mustafaozhan.github.com.mycurrencies.function
+
+inline fun <reified T, reified R> T.mapTo(
+    transform: T.(map: T) -> R?
+): R? =
+    transform(this)
+
+inline fun <reified T, reified R> T.castTo() =
+    if (this != null) {
+        this as? R
+    } else {
+        null
+    }
