+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/function/WhetherScope.kt
@@ -1,4 +1,4 @@
-package mustafaozhan.github.com.mycurrencies.extensions
+package mustafaozhan.github.com.mycurrencies.function
 
 inline fun <reified T> T.whether(
     method: T.(condition: T) -> Boolean
@@ -44,48 +44,6 @@ inline fun <reified T> T.whetherNot(
         } else {
             null
         }
-    } else {
-        null
-    }
-
-inline fun <reified T> T.either(
-    vararg method: T.(condition: T) -> Boolean
-) =
-    if (this != null) {
-        var temp = false
-        method.forEach { temp = temp || it(this) }
-        if (temp) {
-            this
-        } else {
-            null
-        }
-    } else {
-        null
-    }
-
-inline fun <reified T> T.eitherNot(
-    vararg method: T.(condition: T) -> Boolean
-) =
-    if (this != null) {
-        var temp = false
-        method.forEach { temp = temp || it(this) }
-        if (!temp) {
-            this
-        } else {
-            null
-        }
-    } else {
-        null
-    }
-
-inline fun <reified T, reified R> T.mapTo(
-    transform: T.(map: T) -> R?
-): R? =
-    transform(this)
-
-inline fun <reified T, reified R> T.castTo() =
-    if (this != null) {
-        this as? R
     } else {
         null
     }
