+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/ui/main/fragment/calculator/CalculatorViewModel.kt
@@ -10,17 +10,17 @@ import mustafaozhan.github.com.mycurrencies.base.viewmodel.BaseDataViewModel
 import mustafaozhan.github.com.mycurrencies.data.repository.BackendRepository
 import mustafaozhan.github.com.mycurrencies.data.repository.PreferencesRepository
 import mustafaozhan.github.com.mycurrencies.extensions.calculateResult
-import mustafaozhan.github.com.mycurrencies.extensions.either
 import mustafaozhan.github.com.mycurrencies.extensions.getFormatted
 import mustafaozhan.github.com.mycurrencies.extensions.getThroughReflection
 import mustafaozhan.github.com.mycurrencies.extensions.insertInitialCurrencies
-import mustafaozhan.github.com.mycurrencies.extensions.mapTo
 import mustafaozhan.github.com.mycurrencies.extensions.removeUnUsedCurrencies
 import mustafaozhan.github.com.mycurrencies.extensions.replaceNonStandardDigits
 import mustafaozhan.github.com.mycurrencies.extensions.replaceUnsupportedCharacters
 import mustafaozhan.github.com.mycurrencies.extensions.toPercent
-import mustafaozhan.github.com.mycurrencies.extensions.whether
-import mustafaozhan.github.com.mycurrencies.extensions.whetherNot
+import mustafaozhan.github.com.mycurrencies.function.either
+import mustafaozhan.github.com.mycurrencies.function.mapTo
+import mustafaozhan.github.com.mycurrencies.function.whether
+import mustafaozhan.github.com.mycurrencies.function.whetherNot
 import mustafaozhan.github.com.mycurrencies.model.Currencies
 import mustafaozhan.github.com.mycurrencies.model.Currency
 import mustafaozhan.github.com.mycurrencies.model.CurrencyResponse
