+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/ui/main/fragment/settings/SettingsViewModel.kt
@@ -3,9 +3,9 @@ package mustafaozhan.github.com.mycurrencies.ui.main.fragment.settings
 import io.reactivex.Completable
 import mustafaozhan.github.com.mycurrencies.base.viewmodel.BaseDataViewModel
 import mustafaozhan.github.com.mycurrencies.data.repository.PreferencesRepository
-import mustafaozhan.github.com.mycurrencies.extensions.either
 import mustafaozhan.github.com.mycurrencies.extensions.insertInitialCurrencies
 import mustafaozhan.github.com.mycurrencies.extensions.removeUnUsedCurrencies
+import mustafaozhan.github.com.mycurrencies.function.either
 import mustafaozhan.github.com.mycurrencies.model.Currencies
 import mustafaozhan.github.com.mycurrencies.model.Currency
 import mustafaozhan.github.com.mycurrencies.room.dao.CurrencyDao
