+++ app/src/test/java/mustafaozhan/github/com/mycurrencies/function/EitherScopeTest.kt
@@ -1,0 +1,107 @@
+package mustafaozhan.github.com.mycurrencies.function
+
+import mustafaozhan.github.com.mycurrencies.extensions.ScopeFunctionTest
+import org.junit.Assert
+import org.junit.Test
+
+class EitherScopeTest : ScopeFunctionTest() {
+
+    @Test
+    fun `either mix mix`() {
+        subject
+            ?.either({ it.trueCondition }, { it.falseCondition })
+            ?.either({ trueCondition }, { falseCondition })
+            ?.let { Assert.assertTrue(EXPECTED, true) }
+            ?: run { Assert.fail(UN_EXPECTED) }
+
+        subject
+            ?.either({ it.falseCondition }, { it.trueCondition })
+            ?.either({ falseCondition }, { trueCondition })
+            ?.let { Assert.assertTrue(EXPECTED, true) }
+            ?: run { Assert.fail(UN_EXPECTED) }
+    }
+
+    @Test
+    fun `either mix true`() {
+        subject
+            ?.either({ it.trueCondition }, { it.falseCondition })
+            ?.either({ trueCondition }, { trueCondition })
+            ?.let { Assert.assertTrue(EXPECTED, true) }
+            ?: run { Assert.fail(UN_EXPECTED) }
+
+        subject
+            ?.either({ it.falseCondition }, { it.trueCondition })
+            ?.either({ trueCondition }, { trueCondition })
+            ?.let { Assert.assertTrue(EXPECTED, true) }
+            ?: run { Assert.fail(UN_EXPECTED) }
+    }
+
+    @Test
+    fun `either mix false`() {
+        subject
+            ?.either({ it.trueCondition }, { it.falseCondition })
+            ?.either({ falseCondition }, { falseCondition })
+            ?.let { Assert.fail(UN_EXPECTED) }
+            ?: run { Assert.assertTrue(EXPECTED, true) }
+
+        subject
+            ?.either({ it.falseCondition }, { it.trueCondition })
+            ?.either({ falseCondition }, { falseCondition })
+            ?.let { Assert.fail(UN_EXPECTED) }
+            ?: run { Assert.assertTrue(EXPECTED, true) }
+    }
+
+    @Test
+    fun `eitherNot mix`() {
+        subject
+            ?.eitherNot({ it.trueCondition }, { it.falseCondition })
+            ?.eitherNot({ trueCondition }, { falseCondition })
+            ?.let { Assert.fail(UN_EXPECTED) }
+            ?: run { Assert.assertTrue(EXPECTED, true) }
+
+        subject
+            ?.eitherNot({ it.falseCondition }, { it.trueCondition })
+            ?.eitherNot({ falseCondition }, { trueCondition })
+            ?.let { Assert.fail(UN_EXPECTED) }
+            ?: run { Assert.assertTrue(EXPECTED, true) }
+    }
+
+    @Test
+    fun `eitherNot mix true`() {
+        subject
+            ?.eitherNot({ it.trueCondition }, { it.falseCondition })
+            ?.eitherNot({ trueCondition }, { trueCondition })
+            ?.let { Assert.fail(UN_EXPECTED) }
+            ?: run { Assert.assertTrue(EXPECTED, true) }
+
+        subject
+            ?.eitherNot({ it.falseCondition }, { it.trueCondition })
+            ?.eitherNot({ trueCondition }, { trueCondition })
+            ?.let { Assert.fail(UN_EXPECTED) }
+            ?: run { Assert.assertTrue(EXPECTED, true) }
+    }
+
+    @Test
+    fun `eitherNot mix false`() {
+        subject
+            ?.eitherNot({ it.trueCondition }, { it.falseCondition })
+            ?.eitherNot({ falseCondition }, { falseCondition })
+            ?.let { Assert.fail(UN_EXPECTED) }
+            ?: run { Assert.assertTrue(EXPECTED, true) }
+
+        subject
+            ?.eitherNot({ it.falseCondition }, { it.trueCondition })
+            ?.eitherNot({ falseCondition }, { falseCondition })
+            ?.let { Assert.fail(UN_EXPECTED) }
+            ?: run { Assert.assertTrue(EXPECTED, true) }
+    }
+
+    @Test
+    fun `eitherNot false false`() {
+        subject
+            ?.eitherNot({ it.falseCondition }, { it.falseCondition })
+            ?.eitherNot({ falseCondition }, { falseCondition })
+            ?.let { Assert.assertTrue(EXPECTED, true) }
+            ?: run { Assert.fail(UN_EXPECTED) }
+    }
+}
