+++ app/src/test/java/mustafaozhan/github/com/mycurrencies/function/TransferScopeTest.kt
@@ -1,0 +1,76 @@
+package mustafaozhan.github.com.mycurrencies.function
+
+import mustafaozhan.github.com.mycurrencies.extensions.ScopeFunctionTest
+import org.junit.Assert
+import org.junit.Assert.assertEquals
+import org.junit.Assert.assertTrue
+import org.junit.Test
+
+class TransferScopeTest : ScopeFunctionTest() {
+    @Test
+    fun mapTo() {
+        subject
+            ?.mapTo { it.trueCondition }
+            ?.whether { it }
+            ?.let { assertTrue(EXPECTED, true) }
+            ?: run { Assert.fail(UN_EXPECTED) }
+
+        subject
+            ?.mapTo { it.falseCondition }
+            ?.whether { it }
+            ?.let { Assert.fail(UN_EXPECTED) }
+            ?: run { assertTrue(EXPECTED, true) }
+        subject
+            ?.mapTo { trueCondition }
+            ?.whether { it }
+            ?.let { assertTrue(EXPECTED, true) }
+            ?: run { Assert.fail(UN_EXPECTED) }
+
+        subject
+            ?.mapTo { falseCondition }
+            ?.whether { it }
+            ?.let { Assert.fail(UN_EXPECTED) }
+            ?: run { assertTrue(EXPECTED, true) }
+    }
+
+    @Test
+    fun `extraordinary mapTo`() = subject
+        .mapTo { SOME_STRING }
+        ?.mapTo { it -> it.length }
+        ?.let { assertEquals(11, it) }
+        ?: run { Assert.fail(UN_EXPECTED) }
+            .mapTo { subject?.trueCondition }
+            ?.whether { it }
+            ?.mapTo { !it }
+            ?.let { Assert.fail(UN_EXPECTED) }
+        ?: run { assertTrue(EXPECTED, true) }
+
+    @Test
+    fun castTo() {
+        B().castTo<B, A>()
+            ?.let { assertTrue(EXPECTED, true) }
+            ?: run { Assert.fail(UN_EXPECTED) }
+
+        A().castTo<A, B>()
+            ?.let { Assert.fail(UN_EXPECTED) }
+            ?: run { assertTrue(EXPECTED, true) }
+    }
+
+    @Test
+    fun `multi castTo`() {
+        C().castTo<C, B>()
+            ?.castTo<B, A>()
+            ?.let { assertTrue(EXPECTED, true) }
+            ?: run { Assert.fail(UN_EXPECTED) }
+
+        A().castTo<A, B>()
+            ?.castTo<B, C>()
+            ?.let { Assert.fail(UN_EXPECTED) }
+            ?: run { assertTrue(EXPECTED, true) }
+    }
+
+    @Test
+    fun `extraordinary castTo`() = C().castTo<C, A>()
+        ?.let { assertTrue(EXPECTED, true) }
+        ?: run { Assert.fail(UN_EXPECTED) }
+}
